+++ core/common/src/main/kotlin/com/github/jcornaz/miop/Suspension.kt
@@ -1,0 +1,10 @@
+package com.github.jcornaz.miop
+
+import kotlinx.coroutines.suspendCancellableCoroutine
+
+/**
+ * Suspend until the cancellation of the current coroutines.
+ *
+ * Throw a [kotlinx.coroutines.CancellationException] as soon as the coroutines is cancelled
+ */
+public suspend fun awaitCancel(): Nothing = suspendCancellableCoroutine<Nothing> { }