+++ core/common/src/test/kotlin/com/github/jcornaz/miop/experimental/operator/DummyException.kt
@@ -1,3 +1,0 @@
-package com.github.jcornaz.miop.experimental.operator
-
-class DummyException(message: String? = null) : Exception(message)