+++ core/common/src/test/kotlin/com/github/jcornaz/miop/operator/DistinctUtilChanged.kt
@@ -1,0 +1,20 @@
+package com.github.jcornaz.miop.operator
+
+import com.github.jcornaz.miop.distinctUntilChanged
+import com.github.jcornaz.miop.receiveChannelOf
+import com.github.jcornaz.miop.test.runTest
+import kotlinx.coroutines.channels.ReceiveChannel
+import kotlinx.coroutines.channels.toList
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class DistinctUtilChanged : DistinctReferenceUntilChanged() {
+
+    override fun <T> ReceiveChannel<T>.operator(): ReceiveChannel<T> =
+        distinctUntilChanged()
+
+    @Test
+    fun shouldNotEmitLastItemIfEquals() = runTest {
+        assertEquals(listOf("a", "b", "a", "b"), receiveChannelOf("a", "a", "b", "b", "a", "b").operator().toList())
+    }
+}