+++ core/js/src/main/kotlin/com/github/jcornaz/miop/experimental/IoPool.kt
@@ -1,6 +1,0 @@
-package com.github.jcornaz.miop.experimental
-
-import kotlinx.coroutines.experimental.CoroutineDispatcher
-import kotlinx.coroutines.experimental.DefaultDispatcher
-
-actual val IoPool: CoroutineDispatcher get() = DefaultDispatcher