+++ internal/test/js/src/main/kotlin/com/github/jcornaz/miop/internal/test/TestUtils.kt
@@ -1,5 +1,0 @@
-package com.github.jcornaz.miop.internal.test
-
-import kotlinx.coroutines.experimental.promise
-
-actual fun runTest(block: suspend () -> Unit): dynamic = promise { block() }