+++ internal/test/jvm/src/main/kotlin/com/github/jcornaz/miop/internal/test/ActualTestUtils.kt
@@ -1,5 +1,0 @@
-package com.github.jcornaz.miop.internal.test
-
-import kotlinx.coroutines.experimental.runBlocking
-
-actual fun runTest(block: suspend () -> Unit) = runBlocking { block() }